Added searchTest.scala
Added searchTest.scala. The purpose of this program is to determine the best way to traverse a queue. the current issue seems to be that each time we want to go to a specific element we are forced to go through each previous element leading to a very large run time.